requirements/app.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
et-xmlfile==1.1.0
fastapi==0.103.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
openpyxl==3.1.2
packaging==23.2
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.1.0
pluggy==1.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyPDF2==3.0.1
pypdfium2==4.21.0
python-docx==1.0.1
python-http-client==3.3.7
python-magic==0.4.27
python-multipart==0.0.6
reportlab==4.0.6
requests==2.31.0
sendgrid==6.10.0
sniffio==1.3.0
soupsieve==2.5
starkbank-ecdsa==2.2.0
starlette==0.27.0
stripe==6.7.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
mysqlclient==2.2.0
slowapi==0.1.8
email-validator==2.1.0